Generic currency system: spirit as a script-defined currency (#10)
* Design: generic currency system with spirit as a script currency

* Plan: generic currency system with spirit as a script currency

* Add ICurrency and CurrencyHandler registry

* Carry a currency id on item and NPC templates

* Resolve a transaction currency from item then vendor then gold

* Ship gold and credits as built-in currencies

* Widen AttributeSet.SP to long to match HP and MP

* Test SP multiply truncation guard

* Add a vendor test fixture with a message-capturing player

* Route vendor purchases through the currency registry

* Test negative-price purchase refusal

* Route vendor sales through the currency registry

* Test partial-stack gold sale

* Dimensions: register spirit and price dimension items in it

* Spirit: refuse zero-value sales like gold

* Fix comment direction on spirit stamping

+++ Goose/AttributeSet.cs
@@ -13,7 +13,7 @@ namespace Goose
     {
         public long HP { get; set; }
         public long MP { get; set; }
-        public int SP { get; set; }
+        public long SP { get; set; }
         public decimal HPPercentRegen { get; set; }
         public int HPStaticRegen { get; set; }
         public decimal MPPercentRegen { get; set; }
@@ -179,7 +179,7 @@ namespace Goose
 
             temp.HP = (long)Math.Ceiling(a1.HP * multiplier);
             temp.MP = (long)Math.Ceiling(a1.MP * multiplier);
-            temp.SP = (int)Math.Ceiling(a1.SP * multiplier);
+            temp.SP = (long)Math.Ceiling(a1.SP * multiplier);
             temp.HPPercentRegen = a1.HPPercentRegen * multiplier;
             temp.HPStaticRegen = (int)Math.Ceiling(a1.HPStaticRegen * multiplier);
             temp.MPPercentRegen = a1.MPPercentRegen * multiplier;
