+++ Goose/Player.cs
@@ -707,7 +707,7 @@ namespace Goose
             this.BaseStats = new AttributeSet();
             this.BaseStats.HP = Convert.ToInt64(reader["player_hp"]);
             this.BaseStats.MP = Convert.ToInt64(reader["player_mp"]);
-            this.BaseStats.SP = Convert.ToInt32(reader["player_sp"]);
+            this.BaseStats.SP = Convert.ToInt64(reader["player_sp"]);
             this.BaseStats.AC = Convert.ToInt32(reader["stat_ac"]);
             this.BaseStats.Strength = Convert.ToInt32(reader["stat_str"]);
             this.BaseStats.Stamina = Convert.ToInt32(reader["stat_sta"]);
